api/requirements.txt
psycopg2-binary
Flask
flask_restful
flask-cors
cryptography
gunicorn[gevent]
requests